refactor: Merge RequestSendTransactionEvent and RequestSendQubicEvent into RequestSendTransactionEvent
